requirements.txt
flask
python-dotenv
torch
transformers
